Do not use Obj.Extension_constructor.of_val in Type.Id
The `Type.Id.uid` function converts a type identifier to an integer by
going through `Obj.Extension_constructor.of_val`. In this context,
`Obj.Extension_constructor.of_val` is just an expensive way to compute
the identity function: it performs multiple checks, including multiple
calls to `caml_obj_tag`, in order to determine if it is passed a nullary
extension variant, a non-nullary extension variant, or something that is
not an extension variant at all (in which case it raises an
Invalid_argument exception).

When called with a nullary extension variant (which is just represented
as its constructor), `Obj.Extension_constructor.of_val` is always the
identity. Since the usage in `Type.Id` is always with a nullary
extension variant, use `Obj.magic` as a less expensive identity
function.

For context, this came up while optimising code making *many* lookups in
small heterogenous maps, such that it ended up being noticeable in
profiling.